py_scripts/local/requirements.txt
pyzotero==1.6.11
requests==2.32.4
ddgs==9.9.2
duckduckgo-search>=7.0.0
tqdm==4.67.1
beautifulsoup4==4.12.3
python-dotenv==1.1.0
Pillow==11.1.0
google-generativeai
markdown==3.7
